api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg[binary]==3.1.18
alembic==1.13.1
celery==5.3.6
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator==2.1.0.post1
pymupdf==1.23.26
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-multipart==0.0.9
